feat: 项目管理员允许邀请用户，项目非 guest 成员允许添加 ssh 密钥 (#459)

+++ portal/apps/key.go
@@ -3,10 +3,12 @@
 package apps
 
 import (
+	"cloudiac/portal/consts"
 	"cloudiac/portal/consts/e"
 	"cloudiac/portal/libs/ctx"
 	"cloudiac/portal/models"
 	"cloudiac/portal/models/forms"
+	"cloudiac/portal/models/resps"
 	"cloudiac/portal/services"
 	"cloudiac/utils"
 	"fmt"
@@ -25,7 +27,10 @@ func SearchKey(c *ctx.ServiceContext, form *forms.SearchKeyForm) (interface{}, e
 		query = query.Order("created_at DESC")
 	}
 
-	rs, err := getPage(query, form, models.Key{})
+	query = query.
+		Joins("LEFT JOIN iac_user ON iac_user.id = iac_key.creator_id").
+		Select("iac_key.*, iac_user.name AS creator")
+	rs, err := getPage(query, form, resps.KeyResp{})
 	if err != nil {
 		c.Logger().Errorf("error search key, err %s", err)
 		return nil, err
@@ -37,6 +42,16 @@ func SearchKey(c *ctx.ServiceContext, form *forms.SearchKeyForm) (interface{}, e
 // CreateKey 创建密钥
 func CreateKey(c *ctx.ServiceContext, form *forms.CreateKeyForm) (interface{}, e.Error) {
 	c.AddLogField("action", fmt.Sprintf("create key %s", form.Name))
+
+	if !c.IsSuperAdmin && !services.UserHasProjectRole(c.UserId, c.OrgId, c.ProjectId, consts.OrgRoleAdmin) {
+		projectRole, err := services.GetUserHighestProjectRole(c.DB(), c.OrgId, c.UserId)
+		if err != nil {
+			return nil, err
+		} else if projectRole == consts.ProjectRoleGuest {
+			// 除了组织管理员，项目的非 guest 角色也可以添加密钥对，但只能管理自己添加的密钥对
+			return nil, e.New(e.PermissionDeny, http.StatusForbidden)
+		}
+	}
 
 	encrypted, er := utils.AesEncrypt(form.Key)
 	if er != nil {
@@ -69,6 +84,15 @@ func UpdateKey(c *ctx.ServiceContext, form *forms.UpdateKeyForm) (key *models.Ke
 		attrs["name"] = form.Name
 	}
 
+	key = &models.Key{}
+	if err := query.Find(key, form.Id); err != nil {
+		return nil, e.AutoNew(err, e.DBError)
+	}
+	if key.CreatorId != c.UserId && !c.IsSuperAdmin &&
+		!services.UserHasProjectRole(c.UserId, c.OrgId, c.ProjectId, consts.OrgRoleAdmin) {
+		return nil, e.New(e.PermissionDeny, http.StatusOK)
+	}
+
 	key, err = services.UpdateKey(query, form.Id, attrs)
 	if err != nil && (err.Code() == e.KeyAliasDuplicate || err.Code() == e.KeyNotExist) {
 		return nil, e.New(err.Code(), err, http.StatusBadRequest)
@@ -83,6 +107,16 @@ func UpdateKey(c *ctx.ServiceContext, form *forms.UpdateKeyForm) (key *models.Ke
 func DeleteKey(c *ctx.ServiceContext, form *forms.DeleteKeyForm) (result interface{}, re e.Error) {
 	c.AddLogField("action", fmt.Sprintf("delete key %s", form.Id))
 	query := services.QueryKey(services.QueryWithOrgId(c.DB(), c.OrgId))
+
+	key := models.Key{}
+	if err := query.Find(&key, form.Id); err != nil {
+		return nil, e.AutoNew(err, e.DBError)
+	}
+	if key.CreatorId != c.UserId && !c.IsSuperAdmin &&
+		!services.UserHasProjectRole(c.UserId, c.OrgId, c.ProjectId, consts.OrgRoleAdmin) {
+		return nil, e.New(e.PermissionDeny, http.StatusOK)
+	}
+
 	if err := services.DeleteKey(query, form.Id); err != nil {
 		return nil, err
 	}
@@ -92,13 +126,22 @@ func DeleteKey(c *ctx.ServiceContext, form *forms.DeleteKeyForm) (result interfa
 
 func DetailKey(c *ctx.ServiceContext, form *forms.DetailKeyForm) (result interface{}, re e.Error) {
 	query := services.QueryKey(services.QueryWithOrgId(c.DB(), c.OrgId))
-	if key, err := services.GetKeyById(query, form.Id, false); err != nil {
+	key, err := services.GetKeyById(query, form.Id, false)
+	if err != nil {
 		if err.Code() == e.KeyNotExist {
 			return nil, e.New(err.Code(), err, http.StatusBadRequest)
 		}
 		c.Logger().Errorf("error get key by id, err %s", err)
 		return nil, err
-	} else {
-		return key, nil
 	}
+
+	user, err := services.GetUserById(c.DB(), key.CreatorId)
+	if err != nil {
+		return nil, err
+	}
+
+	return resps.KeyResp{
+		Key:     *key,
+		Creator: user.Name,
+	}, nil
 }
